Career Decision Making Agent/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
bcrypt==3.2.2
reportlab==4.2.5
httpx==0.27.2
